requirements.txt
python-dotenv==0.10.1
uvicorn==0.9.0
requests==2.22.0
scikit-learn==0.20.2
dill==0.3.1.1
fastapi==0.41.0
email-validator==1.0.4
jinja2==2.10.3
aiofiles==0.4.0